src/api/requirements.txt
fastapi
fastapi[standard]
requests~=2.32.3
gunicorn==21.2.0
azure-search-documents==11.5.2
azure-keyvault-secrets
aiohttp~=3.11.16
python-dotenv~=1.1.0
azure-cognitiveservices-speech~=1.43.0
jsonlines
nbconvert
pandas
azure-monitor-opentelemetry-exporter~=1.0.0b36
opentelemetry-instrumentation-fastapi~=0.53b1
jupyter
opentelemetry-instrumentation~=0.53b1
azure-identity==1.21.0
azure-ai-projects~=1.0.0b8
gql==3.5.2
openai~=1.72.0